New plugin architecture for Presenter.
To include a plugin, call the use method of the presenter object, passing in the plugin object.
Once the element promise of the presenter is resolved, process will be called on the plugin passing in the root element.
The plugin api is super simple and hopefully can stay that way :)

NB! This is something we might want to expand on in the future, or even remove if we never have the need for plugins. Currently the flyout plugin will be our first use case. We should re-evaluate soon.