Add Flyout presenter plugin.
Will look for all node with data-winning-flyout. For each element, there should be a corresponding flyout component which is passed in to the constructor function.

Flyout.render and show will be called on click of the element, and the flyout will automatically be removed from the dom on hide of the flyout.